requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
openpyxl>=3.1.0
requests>=2.31.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4